+++ CHANGELOG.md
@@ -6,6 +6,7 @@ dev
   in Python 2.6
 - Some minor documentation improvements
 - Exported the `java_timestamp()` function
+- Gave `json2properties` a `--separator` option
 
 v0.1.0 (2016-10-02)
 -------------------